+++ petersburg/__init__.py
@@ -10,6 +10,7 @@
 
 from petersburg.edges import Edge
 from petersburg.estimators import FrequencyEstimator, MixedModeEstimator
+from petersburg.exceptions import PetersburgError, SpecValidationError, ValidationError
 from petersburg.graph import Graph
 from petersburg.nodes import GaussianNode, LogNormalNode, Node, PowerLawNode, UniformNode
 
@@ -23,4 +24,7 @@ __all__ = [
     "Graph",
     "Edge",
     "FrequencyEstimator",
+    "PetersburgError",
+    "ValidationError",
+    "SpecValidationError",
 ]